requirements.txt
requests
pycryptodome==3.19.1
Flask==3.0.0